fix: use name from package json within the patch
To make myself more familiar with wiby I started testing with two repos created for testing purposes.

The [dependency repo my-dependency-id-a](https://github.com/ghinks/my-dependency-id-a) and the [parent consuming it](https://github.com/ghinks/my-parent)

the target modules would be @gvhinks/my-dependency-id-a and the parent @gvhinks/my-parent.

What I found was that the name from the package.json should be used and not the repo name in the patch.