test: remove redundant tests
This is mostly triggered by the upcoming changes in the wiby-test/pass repo, which would require us to update `DEP_HEAD_SHA`. Maintaining this is a low value activity, since we already have unit test coverage for this and we'll soon be able to test integration via the wiby action.

Technically, this is a breaking change, but the exports removed were only ever exposed for testing purposes and were never intended to be used as an API. We're pre-v1 anyways.