Feature/documentation/raison detre (#112)
* feat: raise PR

rationalize repeated nock code block into a beforeEach clause.

* feat: raise PR

Add some text to describe what wiby is as a prelude

* feat: raise PR

add a simple piece of graphics to see if that helps

* docs: readme

- make a simple example with screen shots available
- create some text to give substance to why we have this tool

* docs: readme

- add callouts to cli
- add callouts to PR raised

* docs: readme

- add wiby clean simple example

* docs: readme

- add wiby clean simple example, typo

* docs: readme

- wiby clean, grammar

* docs: readme

- wiby close-pr, interim push to look at github markup

* docs: readme

- wiby close-pr, add image showing PR is closed but branch remains

* docs: readme

- wiby validate , readme

* docs: readme

- wiby validate , typo in the URI

* feat: docs

Add the wiby result command to the readme.

* feat: docs

typos

* feat: docs

add links to the 5 examples given

* feat: docs

add a bit more to explain why we have wiby.

* feat: docs

add links to example repos

* fix: array sz zero handling

the array length needs to be checked as an array that is empty will be returned by the reduce and promise all combination

* fix: undo checkin on incorrect branch

undo checkin on incorrect branch.

* run generate-usage

---------

Co-authored-by: Sebastian Beltran <bjohansebas@gmail.com>